Add sonner and remove unused reducer variable
