+++ components/planner/Board/FilterToolbar/FilterToolbar.tsx
@@ -19,7 +19,7 @@ export const FilterToolbar = ({ boardId }: FilterToolbarProps) => {
         <div className='flex justify-start gap-2 w-full'>
           <TaskFilterSearchBar />
           <DateFilterDisplay />
-          <CategoryFilter />
+          <CategoryFilter selectedBoard={boardId} />
           <ResetButton />
           <AddNewColumnButton key={boardId} boardId={boardId} />
         </div>
