+++ components/planner/Board/TaskColumns/TaskCard/InitializingTaskCard/InitializingTaskCard.tsx
@@ -11,6 +11,7 @@ import { useAuth } from '@clerk/nextjs'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { useState } from 'react'
 import { useForm } from 'react-hook-form'
+import { toast } from 'sonner'
 import * as z from 'zod'
 import { CancelButton } from './CancelButton'
 import { CategoryBadge } from './CategoryBadge'
@@ -64,6 +65,7 @@ export const InitializingTaskCard = ({ boardId, columnId }: InitializingTaskCard
       content: values.taskCardDesc,
     }
     addNewCardToColumn(columns[columnId], newTaskCardDetails, dispatch, getToken)
+    toast.success('Task added.')
   }
 
   return (
