+++ components/planner/Settings/ManageCategoriesCard/AddNewCategoryDialogContent.tsx
