+++ components/planner/Settings/ManageCategoriesCard/CategoryColorPicker.tsx
