+++ components/planner/Settings/ManageCategoriesCard/ManageCategoriesCard.tsx
