+++ components/planner/Settings/ManageColumnsCard/AddNewColumnDialogContent.tsx
