+++ package.json
@@ -53,6 +53,7 @@
     "react-hook-form": "^7.47.0",
     "react-icons": "^4.12.0",
     "react-live-clock": "^6.1.18",
+    "sonner": "^1.5.0",
     "swr": "^2.2.5",
     "tailwind-merge": "^1.14.0",
     "tailwindcss-animate": "^1.0.7",
